Fixed some annoying bugs.  Previously, when you unimported a gnbd (with gnbd_import -R or -r) you could try to kill a gnbd_recvd process that was already dead.. This would result in you killing the pid -1, and bad things happening.
Also, gnbd_monitor wasn't accepting connections, which was causing commands
to hang. This problem was introduced in fixing an earlier bug.

Both these issues are fixed.

+++ gnbd/tools/gnbd_import/gnbd_import.c
@@ -500,7 +500,8 @@ void remove_gnbd(char *name, int minor, int pid)
              name);
     }
   }
-  kill(pid, SIGKILL);
+  if (pid > 0)
+  	kill(pid, SIGKILL);
   cleanup_device(name, minor, fd);
   close(fd);
 }
